feat:TOP-58 DailyTopicUiModel Mapper 구현
 - key를 Long에서 Int타입으로 변경

+++ feature/tohot/src/main/java/tht/feature/tohot/model/TopicUiModel.kt
@@ -8,7 +8,7 @@ import tht.feature.tohot.R
 data class TopicUiModel(
     val imageUrl: String?,
     @DrawableRes val imageRes: Int,
-    val key: Long,
+    val key: Int,
     val title: String,
     val content: String,
     val iconUrl: String,
